init 1.1-SNAPSHOT, add GoogleChartDotWriter
