+++ .gitignore
@@ -1,0 +1,6 @@
+target/
+*.iml
+.idea
+.DS_Store
+.java-version
+dependency-reduced-pom.xml
