+++ dot-diagram/pom.xml
@@ -3,11 +3,21 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.livingdocumentation</groupId>
 	<artifactId>dot-diagram</artifactId>
-	<version>1.0</version>
+	<version>1.1-SNAPSHOT</version>
 	<name>Dot Diagram</name>
 	<description>A simple Java wrapper on top of Graphviz dot</description>
 
+	<properties>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
+	</properties>
+
 	<dependencies>
+		<dependency>
+			<groupId>com.github.kevinsawicki</groupId>
+			<artifactId>http-request</artifactId>
+			<version>5.5</version>
+		</dependency>
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
