+++ dot-diagram/src/main/resources/org/livingdocumentation/dotdiagram/dotstyles.properties
