+++ dot-diagram/src/test/resources/org/livingdocumentation/dotdiagram/graphviz-dot.properties
@@ -1,6 +1,6 @@
 
 # The path where to output generated documentation, with ending slash
-outpath=
+outpath=target/
 
 # The path to the dot tool (Graphviz), with ending slash
 dotpath=/usr/local/bin/
